feat: No more default provider label (#42)
Removed the concept of default provider label as it was confusing and easily caused runners to be left behind. For example, if the user had a typo in the label on the workflow side, the default provider would spin up and never start the job because the labels don't match.

BREAKING CHANGE: `GitHubRunners` no longer takes `defaultProviderLabel` and all workflows must always specify the label they need instead of just `self-hosted`