fix: Try running EC2 instances in all selected subnets (#216)
Previously we only used the first subnet due to a limitation of ec2:RunInstances. This fix adds step function states to try all subnets one by one until one works. This should avoid issues where some instance types are not available on certain availability zones. Unlike the old way of not specifying a subnet, this should work for both the default VPC and a custom VPC.

We use Parallel block (with one branch so not really parallel) to enable retrying of the whole thing when all subnets fail.

Fixes #214