feat: Add basic metrics (#224)
Mostly we expose the built-in metrics of the step function, but the big thing we add is the jobs completed metric. It looks into the runner logs and extracts whether the job itself was successful. This is done using metric filters. The status of the job is exported as a dimension. The provider labels are also exported as a dimension so that data can be easily sliced.

To make the step function based metrics more meaningful, we no longer start it for non self-hosted runs. We check the self-hosted label in the webhook handler now instead of as the first step of the step function.

Fixes #59 

BREAKING CHANGE: IRunnerProvider has a new logGroup field