feat: Always use latest base AMI (#290)
Previously we used SSM to get the latest AMI version during deployment time. This change switches to using AWS Image Builder image ARN to specify the base AMI. When using an image ARN we can either point to a specific image version, or point to the latest image by not specifying a version. This way we will always get the latest base AMI even on weekly scheduled rebuilds. There is no more need to redeploy the stack to get the latest AMI with all the recent security updates.

Another benefit of avoiding SSM is that deploys don't randomly build new runner images even when the diff showed no changes.

Finally, we switch the base Windows AMI from container-optimized to standard to fix #252