fix: Delete unused Image Builder version builds (#303)
These builds are not used for anything but a history of runner image builds. But still it's easier to view when there are no old versions around pointing to AMI or Docker images that already expired.

This PR adds a singleton reaper Lambda that operates on a single AWS Image Builder recipe. It looks for all image versions for that given recipe. For each version, it looks at each version build. For each build, it looks at all the linked AMIs and Docker images. If they are all gone or not the latest, it deletes the version build.

A schedule is created for each recipe to run the reaper Lambda every day.

Fixes #294