fix: Lambda image unable to find runner.js (#307)
The code was trying to load an asset relative to the source file, but assets are always located in the `assets` folder. This only worked in development because the assets are still next to the source code while developing. All assets must always be loaded from the `assets` folder.

Fixes #305