feat: Retry until the job times out (#329)
With proper retry we can better avoid jobs waiting forever for a runner that failed, and stolen runners caused by previous runners failing to start.

GitHub jobs timeout after 24 hours, so we don't have to retry forever. But the old retry of 1 hour wasn't enough.

We still only retry on what we define as retryable errors. The idea is not to retry on obvious configuration issues like a missing AMI or something that causes the runner process itself to start like an old version the server doesn't support. This way we don't waste money on restarting the infrastructure over and over again.

BREAKING CHANGE: `IRunnerProvider` has a new `retryableErrors` required field