fix: Storage setting on EC2/ECS We assumed the root device name was `/dev/sda1`, but that's not always the case. It was not true for all the default ECS AMIs. Those use `/dev/xvda`. That made setting storage size on ECS fail. It could also possibly fail when using different base AMIs for EC2 that use `/dev/xvda`.
This fix uses a custom resource to determine the root device name of a given AMI. We have to be able to handle the `resolve:ssm:*` syntax used by ECS, and passing a launch template as that's what's used by EC2. The custom resource will find the AMI from SSM or the launch template and then get its root device name.