feat: More aggressive retries Try harder to not let a job stay pending. Retry whenever a runner fails for any reason.
The only gotcha here is idle runners. We don't want to let runners wait forever for a job in case it was cancelled. That's a waste of money. Our way of stopping idle runners by deleting them causes the runner process to fail. This bubbles up to the step function with a generic error. We won't be able to filter out that error in the step function to avoid endless retries on idle runners.

Instead of trying to figure out how to get the step function to recognize an idle runner being deleted situation, we just stop the step function too. This has the added benefit of marking it very clearly as an idle runner situation. The step function has a special aborted state and the error code is `IdleRunner`.

Having a special aborted state also helps us monitor the step function for real failures by only looking for failed step functions. Idle runners will be marked as aborted and not caught in the failed step function metric.