feat: More aggressive retries (#349)
Try harder to not let a job stay pending. Retry whenever a runner fails for any reason.

The only gotcha here is idle runners. We don't want to let runners wait forever for a job in case it was cancelled. That's a waste of money. Our way of stopping idle runners by deleting them causes the runner process to fail. This bubbles up to the step function with a generic error. We won't be able to filter out that error in the step function to avoid endless retries on idle runners.

```
2023-05-20T22:23:33.612Z	72ef520c-0b31-4343-bba8-11057e3f860d	ERROR	An error occurred: Access denied. System:ServiceIdentity;DDDDDDDD-DDDD-DDDD-DDDD-DDDDDDDDDDDD needs View permissions to perform the action.
2023-05-20T22:23:33.619Z	72ef520c-0b31-4343-bba8-11057e3f860d	INFO	Runner listener exit with retryable error, re-launch runner in 5 seconds.
2023-05-20T22:23:38.002Z	72ef520c-0b31-4343-bba8-11057e3f860d	INFO	Restarting runner...
2023-05-20T22:23:39.574Z	72ef520c-0b31-4343-bba8-11057e3f860d	INFO	√ Connected to GitHub
2023-05-20T22:23:39.738Z	72ef520c-0b31-4343-bba8-11057e3f860d	ERROR	Failed to create a session. The runner registration has been deleted from the server, please re-configure.
2023-05-20T22:23:39.746Z	72ef520c-0b31-4343-bba8-11057e3f860d	INFO	Runner listener exit with terminated error, stop the service, no retry needed.
2023-05-20T22:23:39.746Z	72ef520c-0b31-4343-bba8-11057e3f860d	INFO	Exiting runner...
2023-05-20T22:23:39.746Z	72ef520c-0b31-4343-bba8-11057e3f860d	INFO	Run done
2023-05-20T22:23:39.748Z	72ef520c-0b31-4343-bba8-11057e3f860d	ERROR	Invoke Error 	{
    "errorType": "Error",
    "errorMessage": "Runner failed with exit code 1",
    "stack": [
        "Error: Runner failed with exit code 1",
        "    at ChildProcess.<anonymous> (/var/task/runner.js:24:16)",
        "    at ChildProcess.emit (events.js:400:28)",
        "    at Process.ChildProcess._handle.onexit (internal/child_process.js:285:12)"
    ]
}
```

Instead of trying to figure out how to get the step function to recognize an idle runner being deleted situation, we just stop the step function too. This has the added benefit of marking it very clearly as an idle runner situation. The step function has a special aborted state and the error code is `IdleRunner`.

![image](https://github.com/CloudSnorkel/cdk-github-runners/assets/1156773/47eb498e-51e1-4326-9c97-c49382b206b4)

Having a special aborted state also helps us monitor the step function for real failures by only looking for failed step functions. Idle runners will be marked as aborted and not caught in the failed step function metric.