fix: Delete runner before retrying (#351)
We need to be more proactive about deleting runners on retries, so we don't get errors like:

```
ERROR A runner exists with the same name CloudSnorkel-ghr-tests-16bddd60-f766-11ed-8185-e6d7cc9e0768.
```

This can happen if the runner already started but failed for some reason. Lambda timeout or a random crash can cause this. If we want to retry, we must remove the runner first.