feat: Enable command execution in ECS, Fargate and EC2 (#361)
Let users execute a command inside the runner containers for debugging. This uses SSM Session Manager. For EC2 it's a button in the console. For ECS/Fargate it's the `aws ecs execute-command` CLI.

We can't enable it for CodeBuild because that requires using batch builds. The required code was left commented out.

This also reduces the permissions given to runners for SSM Session Manager. Instead of using the `AmazonSSMManagedInstanceCore` managed policy, we supply only the bare minimum of permissions. We no longer give everything `ssm:GetParameters` on `*` that comes with `AmazonSSMManagedInstanceCore`.

Fixes #101