fix: Don't start two runners with GitHub environments/deployments (#416)
When GitHub environments are used, we get the `queued` webhook twice. Once when the job is first queued and another when the job is approved for deployment into the specific environment. This causes two runners to be started for the same job. This wastes resources.

To fix this, we check `dpeloyment.statuses_url` in the payload. If it exists, it contains a URL with details about the deployment status. When the deployment is still pending, the first item in the array will have `status: 'waiting'`. Once the deployment is approved and runners can actually run it, the first item in the array will have `status: 'queued'`. So if the status of the deployment is `waiting`, we don't start a runner. When the second webhook comes in, the status will be `queued` and we can start a runner.

Fixes #380