fix: Missing permission for new double runner protection with GitHub environments (#427)
We added a workaround for double runners when using GitHub deployments/environments with #416. For private repositories using deployments/environments, this requires a new permission for the GitHub app or personal access token. The new required permission is [read-only for deployments](https://docs.github.com/en/rest/overview/permissions-required-for-github-apps?apiVersion=2022-11-28#repository-permissions-for-deployments).

This PR updates the documentation and the automated setup process to reflect this new requirement. It also makes this new required permission optional for better backwards compatibility with existing installations.

If you need to add this permission to an existing installation, follow these steps:
1. Generate a new `status.json` using something like `aws --region us-east-1 lambda invoke --function-name github-runners-status12345-xyz status.json`. The exact command can be found in the output section of your deployed CloudFormation stack.
2. In `status.json` find your GitHub application URL under `github.auth.app.url` and navigate to it.
3. Click "App settings" on the right
4. Click "Permissions & events" on the left
5. Click to expand "Repository permissions"
6. Scroll down to Deployments and choose Read-only access.
7. Scroll to the bottom and hit "Save changes"
8. Check your email for notification that the permissions have changed and need to be updated for each app installation

You can also choose to delete the app, put a new value in the setup secret (`status.json` -- `github.setup.secretUrl`), and run through the setup process again.

Related to #380

Fixes #426