fix: Remove pinned Docker-in-Docker version (#431)
The upstream standard docker installation appears to support docker in docker now.  Remove the pinned version and use the same version of docker in all cases.  

**The Integration Test has not been run**.  I did run `integ:default:snapshot`.

I made the assumption that this version was pinned because at the time the version in the upstream repo didn't cleanly support docker in docker, and that in the time since the version was pinned the upstream repo had been updated.  This appears to be the case, at least for Ubuntu + x86 + codebuild.  I'm hoping that pre-merge integration tests will validate the remaining use cases are sane.

The version of docker being deployed for Docker in Docker contains a suite of security vulnerabilities that make corporate vulnerability scanners angry (like mine! that's why I'm here).  I'm hoping that the old version of docker can just be removed so that everyone can benefit from being more up to date and having less to deal with from angry security scanning software.

This passes `npm run build` on my system up until the point where it tried to create bundles for all the different language bundles, because I don't have all the toolchains installed.  

I did validate Codebuild + x86_64 + ubuntu, and will be testing codebuild + arm + ubuntu shortly.  I tested with the following:

``` 
   const tweakedImageBuilder = CodeBuildRunnerProvider.imageBuilder(this, 'tweakedImage', {
      architecture: Architecture.X86_64,
      os: Os.LINUX_UBUNTU,
      vpc: props.vpc,
    });
    tweakedImageBuilder.removeComponent(RunnerImageComponent.dockerInDocker());
    tweakedImageBuilder.addComponent(RunnerImageComponent.docker());
```

Followed up by running `docker run hello-world` in a container running in codebuild.