feat: Faster Fargate startups with SOCI (#499)
Build and push SOCI indexes so Fargate can start our runners faster.

Technically only Fargate supports it now. But it's easier to just always build it. Plus ECS and Lambda may support it in the future. That's why we build it for all providers.

This commit also updates CodeBuild image builder to Amazon Linux 2023. Prior to this we were on Ubuntu for x64 and Amazon Linux 2 for arm64. Previous iterations of this PR required this for containerd, but the current one doesn't require it. Still I kept it here to keep the image build environment consistent between platforms.

Resolves #389
Relates #438