fix: Support longer runner image build timeouts (#543)
Support setting longer timeouts for image runner builds. Previously we used custom resources that are limited to 1 hour. Now we will use a wait handle with support for up to 8 hours (the maximum CodeBuild supports). We now also confirm the requested timeout is supported and fail early if it doesn't.

This doesn't affect:
* AMI builds
* Windows Docker image builds
* Explicit AWS Image Builder builds (`builderType: RunnerImageBuilderType.AWS_IMAGE_BUILDER`)

Don't forget to set the builder timeout with:

```
const fargateX64Builder = FargateRunnerProvider.imageBuilder(stack, 'Fargate builder', {
  codeBuildOptions: {
    timeout: cdk.Duration.hours(8), // 8 hours is the max supported by CodeBuild
  },
});
```

Fixes #540