feat: Tag resources with provider name (#582)
Resources belonging to providers will be tagged with `GitHubRunners:Provider` containing the CDK node path which includes the provider id. This should make it possible to see what resources are being used for which provider. The resources provisioned for runners by the provider should also be tagged where possible.

For example:
```
{
  "Key": "GitHubRunners:Provider",
  "Value": "github-runners-test/CodeBuildx64"
}
```