fix: EC2 provider could fail to auto-update runner version (#632)
The provider failed to read the `RUNNER_VERSION` file and assumed the user didn't use `RunnerVersion.latest()`. It then passed `--disableupdate` to `config.sh` which disabled auto-updates. The default configuration updates the image once a week, so this shouldn't have resulted in using a version that's too old. But we still want to make sure we get the latest version when we ask for it.