feat: Runner group option (#630)
Allow setting [runner group](https://docs.github.com/en/actions/hosting-your-own-runners/managing-self-hosted-runners/managing-access-to-self-hosted-runners-using-groups). GitHub Teams and Enterprise allow you to create runner groups and assign them to specific repos. If a repo is not assigned to the group, it can't use it. This allows you to limit certain runners to certain repos.

```
new CodeBuildRunnerProvider(this, 'limited runner', {
   labels: ['special-runner'],
   group: 'special',
});
```

The group must exist or the runner will not start. The following error will show up in the logs:

```
Could not find any self-hosted runner group named "special".
```

Users will still be able to trigger this runner with the correct labels. But the runner will only be able to run jobs from repos allowed to use the group.

Resolves #500