fix: Idle runner reaper can fail with multiple installations (#754)
When the GitHub app is installed on multiple repos and/or organization, they can each get a different installation id. Idle reaper handler kept one Octokit instance with the first installation id it found in its input. Octokit cannot be shared. We need a different one for each installation id. That means that if multiple jobs came in around the same time, idle reaper would fail to get data on some of them.

This issue could lead to runners being left behind when the app is installed on multiple repos.