fix: Lambda runners failing since 2025/08/17 (#770)
We used to build SOCI for every Docker container, just in case the service using it will start supporting SOCI. On 2025/08/17 we released a new version of standalone-soci-indexer that builds SOCI v2 instead of SOCI v1. This is due to AWS deprecating v1.

Lambda runners stopped working because Lambda doesn't support SOCI v2. Instead of having an index in the same registry, SOCI v2 replaces that main image tag with an image index. Lambda doesn't like that and fails to start.

This fix picks and chooses which providers build and push SOCI index. Instead of always creating an index by default, runner images bound to Lambda providers will not create an index.

Fixes #768