fix: Skip CloudWatch Agent for non-EC2 images (#792)
There is a bug with the latest CloudWatch Agent Debian package that prevents it from being installed inside a Docker image.

We don't really need the agent inside a container. We only need it in EC2 runners. This fix skips agent installation completely on non-EC2 runner images to avoid the issue.

https://github.com/aws/amazon-cloudwatch-agent/issues/1895

Fixes #790