fix: Updating asset content doesn't trigger a new runner image build (#821)
When an asset file changes on disk and a new deploy is started, the runner image will not be rebuilt until its build schedule hits.

This PR fixes the issue by including all asset hashes in the calculations used to decide whether a new image is needed or not.

Fixes #819